Fix finding first repeatable ancestor (TEDEFO-2245)
As we go up the tree, stop whenever we find the first repeatable
ancestor instead of going all the way to the root. Fix is for nodes and
notice type content.

Add unit tests for the 2 getFirstRepeatableAncestor() methods, those
would fail before this fix.
Also change invalid test data to cover this, by moving a field above its
repeatable direct parent to a repeatable node above it.