Refactor Label and Translation classes
A Label instance now has all the information for a specific label text:
id, language and text.
The Translation class now has a map of label ids to Label instances.

This does not impact existing drools files, as they are only using
"/labels[ id == $foo ]", which still works as before.