XmlSchemaValidator: Fix repeatability check for multiple steps
As indicated in https://docs.ted.europa.eu/eforms/latest/guide/xml-generation.html
when a field or node is repeatable, it's always the element of the first
step that can be repeated.
So adjust the check accordingly.

This fixes false errors when looking for the last step, and finding
another element with the same name but at another location under the
parent.

Add such a case in the "valid" unit test.