fields: Improve check of absolute XPath uniqueness (TEDEFO-2423)
Also check that no 2 fields or 2 nodes have the same absolute XPath.

Augment unit test data to demonstrate those issues.

We use "<" to ensure the ids of the 2 fields and nodes are distincts
instead of "!=", to avoid having 2 errors for each problem.

For fields, we ignore OPA-* fields, as they can have the same XPath as
another field.