Refactor to have more consistent validator classes
As we have now 3 other Validator implementation in addition to the main
one that uses Drools, they should be more consistent, and grouped
together into their own package.

Increase the scope of the Validator interface, so that it can be used to
execute all validations.

This will also make it easier to add more Validator implementations in
the future.