notice-types: Fix incorrect warnings on BT-735-LotResult
The check on all repeatable ancestors of a field was incorrect, and
would confuse a field identifier with another field identifier that is a
substring of it (BT-735-LotResult and BT-735-Lot).

Use equality operator instead of memberOf to avoid this.

Also update test data, to have this situation and detect incorrect
warnings.