Check that schematron files are well formed and valid (TEDEFO-2929)
Use ph-schematron to validate the Schematron files. This checks that
they are well-formed XML , and validates against the RELAX NG schema
from the ISO standard.

Adding the dependency on ph-schematron seems to change the JAXB
dependencies, so some dependencies have to be added explicitly in
pom.xml to avoid runtime errors about classes not found.