Add new command to run benchmark of Schematron rules (TEDEFO-3637)
Use JMH to run the Schematron rules of the SDK being analyzed against
a set of large XML notices, and write the results to a JSON file.
This allows detecting when a change in the rules makes them slower, by
comparing with results from another run.

The notices used in the benchmark are included as resources, as we need
them to be always the same, and to be large enough to make slowdowns
noticeable. So we can't use the notice examples from the SDK being
analyzed.