Add check on labels references in view templates index (TEDEFO-3877)
Load the content of view-templates.json in a new Fact class.
Check that the label identifiers it contains correspond to existing
labels.

Add new method in SdkValidationSteps to load the view templates index
file. Rename the existing method to clarify it load the view templates,
and adjust existing test feature accordingly.